% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{dabtab.data}
\alias{dabtab.data}
\title{Launch the dabtab.basic app in the default web browser}
\usage{
dabtab.data(state, ...)
}
\arguments{
\item{state}{Path to statefile to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch the dabtab.basic app in the default web browser
}
\examples{
\dontrun{
dabtab.basic()
dabtab.basic("https://github.com/dabtab-rstats/docs/raw/gh-pages/examples/demo-dvd-rnd.state.rda")
dabtab.basic("viewer")
}
}
