% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{dabtab.data_url}
\alias{dabtab.data_url}
\title{Start dabtab.basic app but do not open a browser}
\usage{
dabtab.data_url(state, ...)
}
\arguments{
\item{state}{Path to statefile to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Start dabtab.basic app but do not open a browser
}
\examples{
\dontrun{
dabtab.basic_url()
}
}
