% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{deregister}
\alias{deregister}
\title{Deregister a data.frame or list in dabtab}
\usage{
deregister(
  dataset,
  shiny = shiny::getDefaultReactiveDomain(),
  envir = r_data,
  info = r_info
)
}
\arguments{
\item{dataset}{String containing the name of the data.frame to deregister}

\item{shiny}{Check if function is called from a shiny application}

\item{envir}{Environment to remove data from}

\item{info}{Reactive list with information about available data in dabtab}
}
\description{
Deregister a data.frame or list in dabtab
}
