% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{fix_names}
\alias{fix_names}
\title{Ensure column names are valid}
\usage{
fix_names(x, lower = FALSE)

fix_names(x, lower = FALSE)
}
\arguments{
\item{x}{Data.frame or vector of (column) names}

\item{lower}{Set letters to lower case (TRUE or FALSE)}
}
\description{
Ensure column names are valid

Ensure column names are valid
}
\details{
Remove symbols, trailing and leading spaces, and convert to valid R column names. Opinionated version of \code{\link{make.names}}

Remove symbols, trailing and leading spaces, and convert to valid R column names. Opinionated version of \code{\link{make.names}}
}
\examples{
fix_names(c(" var-name ", "$amount spent", "100"))
fix_names(c(" var-name ", "$amount spent", "100"))
}
